pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evalforge-runtime"
version = "0.1.0"
description = "Runtime engine for EvalForge generated applications"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Jannis Conen" }]
keywords = ["evalforge", "llm", "ai", "evaluation", "runtime", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

[project.urls]
Homepage = "https://github.com/JannisConen/evalforge-runtime"
Repository = "https://github.com/JannisConen/evalforge-runtime"
Issues = "https://github.com/JannisConen/evalforge-runtime/issues"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "litellm>=1.60.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.18",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
exchange = ["httpx>=0.28.0"]
gmail = ["google-auth>=2.0.0"]
scheduling = ["apscheduler>=3.10.0,<4.0.0"]
azure = ["azure-identity>=1.15.0", "azure-keyvault-secrets>=4.8.0"]
aws = ["boto3>=1.34.0"]
langfuse = ["langfuse>=2.0.0"]
all = [
    "google-auth>=2.0.0",
    "apscheduler>=3.10.0,<4.0.0",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.8.0",
    "boto3>=1.34.0",
    "langfuse>=2.0.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]

[project.scripts]
evalforge-runtime = "evalforge_runtime:main"

[tool.hatch.build.targets.wheel]
packages = ["src/evalforge_runtime"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
